requirements.txt
beautifulsoup4==4.4.1
Django==1.9.7
django-appconf==1.0.2
django-blog-zinnia==0.17
django-contrib-comments==1.7.1
django-mptt==0.8.4
django-tagging==0.4.3
django-xmlrpc==0.1.5
mots-vides==2015.5.11
mysqlclient==1.3.7
Pillow==3.2.0
pkg-resources==0.0.0
pyparsing==2.1.5
pytz==2016.4
regex==2016.6.19
